# backend/openapi/openapi-v0.1.0.yaml
openapi: 3.0.3
info:
  title: Task Tracker API
  version: "0.1.0"
  description: |
    Контракты Backend/Frontend для таск-трекера.
    Документ ориентирован на использование в Swagger UI / Redoc и генерацию SDK.

    Базовый префикс всех эндпоинтов: /api/v1
servers:
  - url: http://localhost:8000/api/v1
    description: Основной сервер API
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [Auth]
      security: []
      summary: Вход (получение JWT-токенов)
      description: |
        Аутентификация пользователя по email/паролю (или прокси к SSO — зависит от реализации).
        Возвращает access_token (для Authorization Bearer) и refresh_token (для продления сессии).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Токены + профиль пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/register:
    post:
      operationId: register
      tags: [Auth]
      security: []
      summary: Регистрация пользователя
      description: |
        Создает нового пользователя в системе.
        Используется при первом входе в систему.
        После регистрации пользователь может выполнить login.

        В MVP регистрация может быть открытой.
        В production возможно ограничение (например, только по корпоративной почте).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Пользователь с таким email уже существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [Auth]
      security: []
      summary: Обновление access_token по refresh_token
      description: |
        Используется, когда access_token истёк.
        Refresh-токен должен быть действующим (не отозван и не истёк).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Новый access_token (+ возможно новый refresh_token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /me:
    get:
      operationId: getMe
      tags: [Auth]
      summary: Текущий пользователь
      description: |
        Возвращает профиль пользователя по access_token.
        Полезно для инициализации фронтенда после перезагрузки страницы.
      responses:
        "200":
          description: Профиль пользователя
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /issues/{issueId}/attachments/prepare:
    post:
      operationId: prepareAttachment
      tags: [Attachments]
      summary: Подготовить загрузку вложения (presigned upload)
      parameters:
        - $ref: "#/components/parameters/IssueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrepareAttachmentRequest"
      responses:
        "200":
          description: Данные для загрузки
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrepareAttachmentResponse"
  /issues/{issueId}/attachments:
    get:
      operationId: listAttachments
      tags: [Attachments]
      summary: Список вложений задачи
      parameters:
        - $ref: "#/components/parameters/IssueId"
      responses:
        "200":
          description: Вложения
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/IssueAttachment"
    post:
      operationId: createAttachment
      tags: [Attachments]
      summary: Подтвердить вложение после загрузки
      parameters:
        - $ref: "#/components/parameters/IssueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttachmentRequest"
      responses:
        "201":
          description: Вложение зафиксировано
          content:
            application/json:
              schema:
                type: object
                required: [attachment]
                properties:
                  attachment:
                    $ref: "#/components/schemas/IssueAttachment"
  /attachments/{attachmentId}/download:
    get:
      operationId: getAttachmentDownloadUrl
      tags: [Attachments]
      summary: Получить URL для скачивания
      parameters:
        - $ref: "#/components/parameters/AttachmentId"
      responses:
        "200":
          description: URL для скачивания
          content:
            application/json:
              schema:
                type: object
                required: [download_url]
                properties:
                  download_url:
                    type: string
                    format: uri
  /attachments/{attachmentId}:
    delete:
      operationId: deleteAttachment
      tags: [Attachments]
      summary: Удалить вложение
      parameters:
        - $ref: "#/components/parameters/AttachmentId"
      responses:
        "204":
          description: Удалено
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token в заголовке Authorization: Bearer <token>"
  parameters:
...
    IssueId:
      name: issueId
      in: path
      required: true
      description: UUID задачи/эпика.
      schema:
        $ref: "#/components/schemas/UUID"
...
    AttachmentId:
      name: attachmentId
      in: path
      required: true
      description: UUID вложения.
      schema:
        $ref: "#/components/schemas/UUID"
...
  responses:
    BadRequest:
      description: Ошибка запроса / ошибка валидации
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Нет прав
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    UUID:
      type: string
      format: uuid
      description: UUID v4.
      example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
...
    DateTime:
      type: string
      format: date-time
      description: ISO 8601 datetime (с таймзоной).
      example: "2026-03-02T10:15:30Z"
...
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
          description: Поле запроса.
        reason:
          type: string
          description: Причина.
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Код ошибки.
          example: VALIDATION_ERROR
        message:
          type: string
          description: Текст ошибки.
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
        request_id:
          type: string
          description: ID запроса.
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
    RegisterRequest:
      type: object
      required: [email, full_name, password]
      properties:
        email:
          type: string
          format: email
          description: Email пользователя (логин).
        full_name:
          type: string
          description: Полное имя пользователя.
          example: Иван Иванов
        password:
          type: string
          format: password
          minLength: 8
          description: Пароль пользователя.
    RegisterResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          description: Email пользователя.
        password:
          type: string
          format: password
          description: Пароль.
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          description: Refresh token.
    LoginResponse:
      type: object
      required: [access_token, refresh_token, user]
      properties:
        access_token:
          type: string
          description: Access token.
        refresh_token:
          type: string
          description: Refresh token.
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, full_name, is_active, created_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        email:
          type: string
          format: email
          description: Email.
        full_name:
          type: string
          description: ФИО.
        is_active:
          type: boolean
          description: Активен ли.
        created_at:
          $ref: "#/components/schemas/DateTime"
...
    PrepareAttachmentRequest:
      type: object
      required: [file_name, mime_type, size_bytes]
      properties:
        file_name:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
          minimum: 1
    PrepareAttachmentResponse:
      type: object
      required: [upload]
      properties:
        upload:
          type: object
          required: [storage_key, upload_url, headers, fields, method, expires_in]
          properties:
            storage_key:
              type: string
            upload_url:
              type: string
              format: uri
            headers:
              type: object
              additionalProperties:
                type: string
            fields:
              type: object
              additionalProperties:
                type: string
            method:
              type: string
              enum: [POST]
            expires_in:
              type: integer
    CreateAttachmentRequest:
      type: object
      required: [storage_key, file_name, mime_type, size_bytes]
      properties:
        storage_key:
          type: string
        file_name:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
    IssueAttachment:
      type: object
      required: [id, issue_id, uploaded_by, file_name, mime_type, size_bytes, storage_key, created_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        issue_id:
          $ref: "#/components/schemas/UUID"
        uploaded_by:
          $ref: "#/components/schemas/UUID"
        file_name:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
        storage_key:
          type: string
        created_at:
          $ref: "#/components/schemas/DateTime"
...